Cambio a DTO y metodos HTTP corregidos Personal Bancario

+++ src/main/java/com/banquito/core/banking/seguridadbanco/dao/RolRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.banquito.core.banking.seguridadbanco.domain.Rol;
 
 @Repository
-public interface RolRepository extends CrudRepository<Rol, BigDecimal> {
+public interface RolRepository extends CrudRepository<Rol, Integer> {
     public List<Rol> findByCodRolOrderByCodRol(BigDecimal codRol);
     public List<Rol> findByNombreRolAndResponsableOrderByFechaCreacion(String nombreRol, String responsable);
     public List<Rol> findByCodRolAndResponsableOrderByNombreRol(BigDecimal codRol, String responsable);
